+++ app/Http/Controllers/UserController.php
@@ -6,23 +6,30 @@ use App\Http\Requests\StoreUserRequest;
 use App\Http\Requests\UpdateUserRequest;
 use App\Http\Resources\UserCrudResource;
 use App\Models\User;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class UserController extends Controller
 {
+    use AuthorizesRequests;
+
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        $query = User::query();
+        $user = Auth::user();
+        $workspace = $user->workspaces->first();
 
-        $sortField = request("sort_field", "created_at");
-        $sortDirection = request("sort_direction", "desc");
-        
+        $query = User::inWorkspace($workspace);
+
+        $sortField = request("sort_field", "id");
+        $sortDirection = request("sort_direction", "asc");
+
         if (request("name")) {
             $query->where("name", "like", "%" . request("name") . "%");
         }
-         if (request("email")) {
+        if (request("email")) {
             $query->where("email", "like", "%" . request("email") . "%");
         }
 
@@ -30,10 +37,16 @@ class UserController extends Controller
             ->paginate(10)
             ->onEachSide(1);
 
+        if (request()->page && request()->page > $users->lastPage()) {
+            abort(404);
+        }
+
         return inertia("User/Index", [
             "users" => UserCrudResource::collection($users),
             'queryParams' => request()->query() ?: null,
             'success' => session('success'),
+            'isOwner' => $user->ownedWorkspace !== null,
+            "ownerId" => $user->id,
         ]);
     }
 
@@ -42,6 +55,8 @@ class UserController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', User::class);
+
         return inertia("User/Create");
     }
 
@@ -50,22 +65,22 @@ class UserController extends Controller
      */
     public function store(StoreUserRequest $request)
     {
+        $this->authorize('create', User::class);
+
+        $authUser = Auth::user();
+
         $data = $request->validated();
         $data['email_verified_at'] = time();
         $data['password'] = bcrypt($data['password']);
+        $data['created_by'] = $authUser->id;
 
-        User::create($data);
+        $newUser = User::create($data);
+
+        $workspace = Auth::user()->workspaces->first();
+        $workspace->members()->attach($newUser->id, ['role' => 'member']);
 
         return to_route('user.index')
-            ->with('success', 'User was created');
-    }
-
-    /**
-     * Display the specified resource.
-     */
-    public function show(User $user)
-    {
-        //
+            ->with('success', 'Member created successfully');
     }
 
     /**
@@ -73,6 +88,8 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        $this->authorize('manage', $user);
+
         return inertia("User/Edit", [
             'user' => new UserCrudResource($user),
         ]);
@@ -83,17 +100,20 @@ class UserController extends Controller
      */
     public function update(UpdateUserRequest $request, User $user)
     {
+        $this->authorize('manage', $user);
+
         $data = $request->validated();
         $password = $data['password'] ?? null;
+
         if ($password) {
             $data['password'] = bcrypt($password);
-        }else {
-            unset($data['password']); 
+        } else {
+            unset($data['password']);
         }
         $user->update($data);
 
         return to_route('user.index')
-            ->with('success', "User \"$user->name\" was updated"); 
+            ->with('success', "Member \"$user->name\" updated successfully");
     }
 
     /**
@@ -101,9 +121,11 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        $this->authorize('manage', $user);
+
         $name = $user->name;
-        $user->delete();    
+        $user->delete();
         return to_route('user.index')
-            ->with('success', "User \"$name\" was deleted");
+            ->with('success', "Member \"$name\" deleted successfully");
     }
 }
